renaming from table to row
