+++ wls-gui-wahllokalsystem/src/components/vorfaelleundvorkommnisse/TheEreignisseRow.vue
